Create a separate module for `units` errors
The custom exceptions and warnings `units` define are now in a separate
module. Most code should import them from the `astropy.units` namespace
(as before). In cases this would cause an import loop they should be
imported directly from the `astropy.units.errors` namespace.
